* commit 'e630d4719337d2b8581b2ba1bfdb8b3a13829dd2': (21 commits)
  conflict merge
  Fixed, GPS was working when app was minimized
  Fixed, GPS was working when app was minimized
  copy array differently
  change according to feedback
  Removed unneccessary comments
  Reservation view
  View
  conflict
  remove unnecessary coments
  minor view changes
  Dynamic views
  works but table spinner duplicated
  Empty view
  Small adjustments in the view
  effect
  testing include xml
  Testing include xml
  show available tables
  Restaurant ID hard coded
  ...

+++ app/src/main/java/de/tum/ecorp/reservationapp/resource/RestaurantResourceAsync.java
@@ -13,6 +13,6 @@ public interface RestaurantResourceAsync {
     void getRestaurantsFilteredAsync(final Task<List<Restaurant>> task, final Map<RestaurantResource.Filter, Object> filters);
     void getRestaurantsBySearchStringAsync(final Task<List<Restaurant>> task, final String searchString);
     void getRestaurantsNearbyAsync(final Task<List<Restaurant>> task, final Location searchLocation, final Double searchRadius);
-    void getRestaurantAsync(final Task<Restaurant> task, Long restaurantId);
+    void getRestaurantAsync(final Task<Restaurant> task, Integer restaurantId);
 
 }
